P2/app/requirements.txt
psycopg2-binary==2.9.1
fastapi
uvicorn
sqlalchemy